aspeed: Add early init for KCS and Snoop
When the eSPI-LPC bridge HW encounters an LPC I/O port
decoding error, the LPC bus will be kept in busy state.
And controllers behind the bridge cannot be enabled
when the LPC bus is busy.

This patch enables the Snoop(80h) and the KCS(CA2h)
controllers at early stage to avoid being not able
to be activated due to the issue above.