aspeed: Add eSPI early init
The Host PCH booting could fail because of that the
BMC-side eSPI channels are not ready. This patch adds
early initialization to the eSPI controller. Thus the
eSPI channel can becomes ready at the early stage.

Note that this merely allows PCH to continue its booting.
The eSPI is not fully functional until the BMC kernel
booting is done.

Signed-off-by: Chia-Wei, Wang <chiawei_wang@aspeedtech.com>