ram:aspeed: add missing handshaking flags.
scu100[7]: indicate the dram is initiailized by the soc.
scu100[12]:indicate the p2a bridge is disable

on ast2600, the p2a bridge is disabled by default so the dram must be
initialized by the BMC SOC.  Setting these flags so that the legacy
bios code won't stall in a infinite loop for polling these flags.

Signed-off-by: Dylan Hung <dylan_hung@aspeedtech.com>
Change-Id: I28a274b24757778b733033dd0c3c6e84f8cd1a55