clk: ast2600: Setup RGMII TX delay according to the DLY32 calculation
There is a duplicate DLY32 delay cell embedded in AST2600 SOC that is
the same with the one used for RGMII clock delay.  The delay time of
each tap delay can be measured by the embedded frequency counters
SCU320 and SCU330.  So we can set up the TX delay according to the
runtime measured value to cover the chip-to-chip delay time variation.

v2: revise coding style

Signed-off-by: Dylan Hung <dylan_hung@aspeedtech.com>
Change-Id: I42c6cf09f36baf9ded22accf476cc93ae4991d86