clk: ast2600: Use a for-loop to search the MII delay value
We used binary-search to search the MII delay but this method is
difficult to handle the corner cases such as invalid lookup table, the
target delay time is out of the table.  Now we use a simple for-loop to
find the delay value, if no proper delay time is found, return -1 to
specific the error condition.

Signed-off-by: Dylan Hung <dylan_hung@aspeedtech.com>
Change-Id: I24ec3403b57113f00517da595fddc3d9870bcb07