as2600/scu: Fix printing of security info
The current code misses initialising the char buffer 'alg' to zero,
causing the sprintf to potentially write past the end of the buffer.

Most of the time strlen happened upon a 0 early in the buffer, and the
resulting string would be constructed in bounds:

 Secure Boot: Mode_2, m��ERSA4096_SHA512

Avoid the issue by not constructing the string in memory. Instead print
it out as the bits are parsed.

Fixes: dd27b24b13d5 ("ARM: Aspeed: update secure boot information")
Signed-off-by: Joel Stanley <joel@jms.id.au>
Link: https://lore.kernel.org/r/20220310045801.183051-1-joel@jms.id.au