spi: aspeed: Use "jedec,spi-nor" compatible to count devices
The SoC device trees use the "jedec,spi-nor" compatible property to
describe the flash devices and the board device trees override this
property with "spi-flash", "sst,w25q256". This comes from the initial
driver which was first written when spi-nor support had not been
merged in U-boot yet.

"jedec,spi-nor" should be preferred since it is used by Linux and the
latest U-boot. Use it to count devices. We will clean up the board
device trees later.

Cc: Chin-Ting Kuo <chin-ting_kuo@aspeedtech.com>
Signed-off-by: Cédric Le Goater <clg@kaod.org>
Reviewed-by: Joel Stanley <joel@jms.id.au>
Link: https://lore.kernel.org/r/20220608061455.365123-1-clg@kaod.org
Signed-off-by: Joel Stanley <joel@jms.id.au>