cmd/otp: Use any MOD_EXP driver
Instead of requesting the ACRY driver specifically, ask for the first
DM device that implements MOD_EXP.

Selecting RSA ensures that one of the MOD_EXP drivers will be built in.
On Aspeed platforms this will be either the ACRY or the software
implementation; Kconfig logic stops both from being built in.

Signed-off-by: Joel Stanley <joel@jms.id.au>